working version of log in waterfall
